move policy methods to policy_linux.go
