+++ npm/pkg/dataplane/policies/policy_linux.go
@@ -1,0 +1,27 @@
+package policies
+
+import "github.com/Azure/azure-container-networking/npm/util"
+
+// returns two booleans indicating whether the network policy has ingress and egress respectively
+func (networkPolicy *NPMNetworkPolicy) hasIngressAndEgress() (hasIngress, hasEgress bool) {
+	hasIngress = false
+	hasEgress = false
+	for _, aclPolicy := range networkPolicy.ACLs {
+		hasIngress = hasIngress || aclPolicy.hasIngress()
+		hasEgress = hasEgress || aclPolicy.hasEgress()
+	}
+	return
+}
+
+func (networkPolicy *NPMNetworkPolicy) egressChainName() string {
+	return networkPolicy.chainName(util.IptablesAzureEgressPolicyChainPrefix)
+}
+
+func (networkPolicy *NPMNetworkPolicy) ingressChainName() string {
+	return networkPolicy.chainName(util.IptablesAzureIngressPolicyChainPrefix)
+}
+
+func (networkPolicy *NPMNetworkPolicy) chainName(prefix string) string {
+	policyHash := util.Hash(networkPolicy.Name) // assuming the name is unique
+	return joinWithDash(prefix, policyHash)
+}
