+++ npm/pkg/dataplane/policies/policymanager_linux.go
@@ -81,30 +81,6 @@ func allChainNames(networkPolicies []*NPMNetworkPolicy) []string {
 	return chainNames
 }
 
-// returns two booleans indicating whether the network policy has ingress and egress respectively
-func (networkPolicy *NPMNetworkPolicy) hasIngressAndEgress() (hasIngress, hasEgress bool) {
-	hasIngress = false
-	hasEgress = false
-	for _, aclPolicy := range networkPolicy.ACLs {
-		hasIngress = hasIngress || aclPolicy.hasIngress()
-		hasEgress = hasEgress || aclPolicy.hasEgress()
-	}
-	return
-}
-
-func (networkPolicy *NPMNetworkPolicy) egressChainName() string {
-	return networkPolicy.chainName(util.IptablesAzureEgressPolicyChainPrefix)
-}
-
-func (networkPolicy *NPMNetworkPolicy) ingressChainName() string {
-	return networkPolicy.chainName(util.IptablesAzureIngressPolicyChainPrefix)
-}
-
-func (networkPolicy *NPMNetworkPolicy) chainName(prefix string) string {
-	policyHash := util.Hash(networkPolicy.Name) // assuming the name is unique
-	return joinWithDash(prefix, policyHash)
-}
-
 func (pMgr *PolicyManager) newCreatorWithChains(chainNames []string) *ioutil.FileCreator {
 	creator := ioutil.NewFileCreator(pMgr.ioShim, maxRetryCount, knownLineErrorPattern, unknownLineErrorPattern) // TODO pass an array instead of this ... thing
